requirements.txt
python-telegram-bot[ext]
anthropic
openai
msal
httpx
fastapi
uvicorn[standard]
python-dotenv
notion-client
pytest
pytest-asyncio
